Dodging Finally Works!!!
Added Dodging and something special for the shooters. Monsters currently have 1/1000 odds to dodge each frame, after which, they will not try to dodge the same projectile again. This also kind of counts as a refactor of the walker archetype as I found that they were being assigned to move about 6 times for arbitrary reasons each frame. This means that they are actually better at tracking down the village too.